requirements.txt
aiofiles
aiohttp
apscheduler
cairosvg
cloudscraper
colour
covid
cowpy
dataclasses
DateTime
emoji
fonttools
geopy
gitpython
glitch_this
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
googletrans==4.0.0-rc1
gtts
hachoir
heroku3
html-telegraph-poster
humanize
IMDbPY
jikanpy
justwatch
lottie
lyricsgenius
markdown
motor
moviepy
nekos.py
opencv-python-headless
Pillow
prettytable
psutil
psycopg2
PyDictionary
pyfiglet
PyGithub
pygments
pylast
pymediainfo
PyMuPDF
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
randomstuff.py
regex
requests
search-engine-parser
selenium
setuptools
ShazamAPI
spamwatch
speedtest-cli
sqlalchemy-json
sqlalchemy==1.3.23
telegraph
git+https://github.com/KENZO-404/Telethon
tgcrypto
tswift
ujson
urlextract
validators
vcsi
wand
wget
git+https://github.com/goldsmith/Wikipedia
youtube-search-python
youtube_dl
pre-commit
pytgcalls
ffmpeg-python
python-arq
Pyrogram
TgCrypto
